+++ src/design/plone/policy/tests/test_search_filters_api.py
@@ -2,6 +2,7 @@
 from design.plone.policy.testing import (
     DESIGN_PLONE_POLICY_API_FUNCTIONAL_TESTING,
 )
+from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
@@ -9,8 +10,8 @@ from plone.app.testing import TEST_USER_ID
 from plone.registry.interfaces import IRegistry
 from plone.restapi.testing import RelativeSession
 from Products.CMFPlone.interfaces import ISearchSchema
-from zope.component import getUtility
 from transaction import commit
+from zope.component import getUtility
 
 import unittest
 
@@ -27,19 +28,20 @@ class SearchFiltersAPITest(unittest.TestCase):
         self.api_session = RelativeSession(self.portal_url)
         self.api_session.headers.update({"Accept": "application/json"})
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+        self.api_session_anon = RelativeSession(self.portal_url)
+        self.api_session_anon.headers.update({"Accept": "application/json"})
 
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
     def tearDown(self):
         self.api_session.close()
+        self.api_session_anon.close()
 
     def test_endpoint_exists(self):
         response = self.api_session.get("/@search-filters")
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(
-            response.headers.get("Content-Type"), "application/json"
-        )
+        self.assertEqual(response.headers.get("Content-Type"), "application/json")
 
     def test_endpoint_return_list_of_sections(self):
         response = self.api_session.get("/@search-filters").json()
@@ -59,6 +61,26 @@ class SearchFiltersAPITest(unittest.TestCase):
 
         self.assertIn("topics", response)
         self.assertEqual(response["topics"], [])
+
+    def test_endpoint_return_list_of_topics_based_on_roles(self):
+
+        api.content.create(
+            container=self.portal, type="Pagina Argomento", title="private topic"
+        )
+        public_topic = api.content.create(
+            container=self.portal, type="Pagina Argomento", title="public topic"
+        )
+        api.content.transition(obj=public_topic, transition="publish")
+        commit()
+
+        response = self.api_session.get("/@search-filters").json()
+        self.assertIn("topics", response)
+        self.assertEqual(len(response["topics"]), 2)
+
+        response = self.api_session_anon.get("/@search-filters").json()
+
+        self.assertIn("topics", response)
+        self.assertEqual(len(response["topics"]), 1)
 
     def test_endpoint_return_list_of_searchable_types(self):
         response = self.api_session.get("/@search-filters").json()